Restructure tests into separate files
