+++ oeplatform/settings.py
@@ -44,6 +44,7 @@ INSTALLED_APPS = (
     'colorfield',
     'django_ajax',
     'literature',
+    'api',
     'axes'
 )
